+++ Garage-Management-Application/src/main/java/com/danny/Garage/Management/Application/dto/LabourUpdateDTO.java
@@ -1,13 +1,15 @@
 package com.danny.Garage.Management.Application.dto;
 
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class LabourUpdateDTO {
     private Long billId;
-    private Long labourAmount;
+    private BigDecimal labourAmount;
 }
